pyproject.toml
[project]
name = "telegram-openai-whisper-api-transcriber"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "aiosqlite>=0.21.0",
    "django>=4.2.25",
    "elevenlabs>=2.16.0",
    "google-genai>=1.45.0",
    "openai>=2.0.1",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
    "sentry-sdk>=2.39.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-django>=4.11.1",
    "ruff>=0.13.3",
]
